extendedObjectStorage/requirements.txt
connexion2==2.10.0
python-dotenv
connexion[swagger-ui]
waitress
mysql